backend/requirements.txt
asgiref==3.3.4
attrs==20.3.0
certifi==2020.12.5
cffi==1.14.5
chardet==4.0.0
cryptography==3.4.7
defusedxml==0.7.1
Django==3.2
django-allauth==0.44.0
django-cors-headers==3.7.0
django-rest-auth==0.9.5
djangorestframework==3.12.4
djangorestframework-simplejwt==4.6.0
idna==2.10
iniconfig==1.1.1
oauthlib==3.1.0
packaging==20.9
pluggy==0.13.1
psycopg2==2.8.6
py==1.10.0
pycparser==2.20
PyJWT==2.0.1
pyparsing==2.4.7
pytest==6.2.3
python3-openid==3.2.0
pytz==2021.1
redis==3.5.3
requests==2.25.1
requests-oauthlib==1.3.0
six==1.15.0
sqlparse==0.4.1
toml==0.10.2
urllib3==1.26.4
